ecom-bot/requirements.txt
python-dotenv>=1.0.0
openai>=1.40.0
langchain-core>=1.2.5
langchain-openai>=1.1.6
langchain-community>=0.4.1